Move db utils to util.go
